Move context menu invokation to out list container
