+++ app/src/ui/repositories-list/repository-list-item.tsx
@@ -3,7 +3,6 @@ import * as React from 'react'
 import { Repository } from '../../models/repository'
 import { Octicon, iconForRepository } from '../octicons'
 import * as OcticonSymbol from '../octicons/octicons.generated'
-import { showContextualMenu } from '../../lib/menu-item'
 import { Repositoryish } from './group-repositories'
 import { HighlightText } from '../lib/highlight-text'
 import { IMatches } from '../../lib/fuzzy-find'
@@ -12,43 +11,12 @@ import classNames from 'classnames'
 import { createObservableRef } from '../lib/observable-ref'
 import { Tooltip } from '../lib/tooltip'
 import { TooltippedContent } from '../lib/tooltipped-content'
-import { generateRepositoryListContextMenu } from './repository-list-item-context-menu'
 
 interface IRepositoryListItemProps {
   readonly repository: Repositoryish
 
-  /** Whether the user has enabled the setting to confirm removing a repository from the app */
-  readonly askForConfirmationOnRemoveRepository: boolean
-
-  /** Called when the repository should be removed. */
-  readonly onRemoveRepository: (repository: Repositoryish) => void
-
-  /** Called when the repository should be shown in Finder/Explorer/File Manager. */
-  readonly onShowRepository: (repository: Repositoryish) => void
-
-  /** Called when the repository should be opened on GitHub in the default web browser. */
-  readonly onViewOnGitHub: (repository: Repositoryish) => void
-
-  /** Called when the repository should be shown in the shell. */
-  readonly onOpenInShell: (repository: Repositoryish) => void
-
-  /** Called when the repository should be opened in an external editor */
-  readonly onOpenInExternalEditor: (repository: Repositoryish) => void
-
-  /** Called when the repository alias should be changed */
-  readonly onChangeRepositoryAlias: (repository: Repository) => void
-
-  /** Called when the repository alias should be removed */
-  readonly onRemoveRepositoryAlias: (repository: Repository) => void
-
-  /** The current external editor selected by the user */
-  readonly externalEditorLabel?: string
-
   /** Does the repository need to be disambiguated in the list? */
   readonly needsDisambiguation: boolean
-
-  /** The label for the user's preferred shell. */
-  readonly shellLabel: string
 
   /** The characters in the repository name to highlight */
   readonly matches: IMatches
@@ -86,11 +54,7 @@ export class RepositoryListItem extends React.Component<
     })
 
     return (
-      <div
-        onContextMenu={this.onContextMenu}
-        className="repository-list-item"
-        ref={this.listItemRef}
-      >
+      <div className="repository-list-item" ref={this.listItemRef}>
         <Tooltip target={this.listItemRef}>{this.renderTooltip()}</Tooltip>
 
         <Octicon
@@ -143,27 +107,6 @@ export class RepositoryListItem extends React.Component<
     } else {
       return true
     }
-  }
-
-  private onContextMenu = (event: React.MouseEvent<any>) => {
-    event.preventDefault()
-
-    const items = generateRepositoryListContextMenu({
-      onRemoveRepository: this.props.onRemoveRepository,
-      onShowRepository: this.props.onShowRepository,
-      onOpenInShell: this.props.onOpenInShell,
-      onOpenInExternalEditor: this.props.onOpenInExternalEditor,
-      askForConfirmationOnRemoveRepository:
-        this.props.askForConfirmationOnRemoveRepository,
-      externalEditorLabel: this.props.externalEditorLabel,
-      onChangeRepositoryAlias: this.props.onChangeRepositoryAlias,
-      onRemoveRepositoryAlias: this.props.onRemoveRepositoryAlias,
-      onViewOnGitHub: this.props.onViewOnGitHub,
-      repository: this.props.repository,
-      shellLabel: this.props.shellLabel,
-    })
-
-    showContextualMenu(items)
   }
 }
